+++ tests/urls.py
@@ -7,7 +7,5 @@ from django.contrib import admin
 
 
 urlpatterns = [
-    path(r'^admin/', admin.site.urls),
+    path('admin/', admin.site.urls),
 ]
-
-